replace some v1 sub-account endpoints with v2
