+++ examples/spot/corporate/sub_account_futures_account_summary.py
@@ -10,4 +10,4 @@ key = ''
 secret = ''
 
 spot_client = Client(key, secret)
-logging.info(spot_client.sub_account_futures_account_summary())
+logging.info(spot_client.sub_account_futures_account_summary(futuresType=1))
