--- examples/spot/corporate/sub_account_futures_position_risk_v2.py
@@ -1,16 +1,0 @@
-#!/usr/bin/env python
-
-import logging
-from binance.spot import Spot as Client
-from binance.lib.utils import config_logging
-
-config_logging(logging, logging.DEBUG)
-
-key = ''
-secret = ''
-
-spot_client = Client(key, secret)
-logging.info(spot_client.sub_account_futures_position_risk_v2(
-    email='alice@test.com',
-    futuresType=1
-))
