+++ tests/spot/corporate/test_sub_account_futures_account_summary.py
@@ -12,10 +12,22 @@ key = random_str()
 secret = random_str()
 
 
-@mock_http_response(responses.GET, '/sapi/v1/sub-account/futures/accountSummary', mock_item, 200)
-def test_sub_account_futures_account_summary():
-    """ Tests the API endpoint to  get sub account futures account summary """
+def test_sub_account_futures_account_summary_v2_without_futuresType():
+    """ Tests the API endpoint to  get sub account futures account summary without futuresType """
+
+    params = {
+        'futuresType': '',
+        'recvWindow': 1000
+    }
 
     client = Client(key, secret)
-    response = client.sub_account_futures_account_summary()
+    client.sub_account_futures_account_summary.when.called_with(**params).should.throw(ParameterRequiredError)
+
+
+@mock_http_response(responses.GET, '/sapi/v2/sub-account/futures/accountSummary\\?futuresType=1', mock_item, 200)
+def test_sub_account_futures_account_summary_v2():
+    """ Tests the API endpoint to get sub account futures account summary """
+
+    client = Client(key, secret)
+    response = client.sub_account_futures_account_summary(futuresType=1)
     response.should.equal(mock_item)
